pal(): handle negative c1 in screen-palette mode
Carts use pal(c0, NEGATIVE, 1) as a HUD-hiding trick — the intent
appears to be "make color c0 invisible in screen output". With the
previous uint8_t parameter type, -15 cast to 0xF1 produced a
screen_palette entry of 241, which after &0x8f became 0x81 and
routed through the extended palette LUT to display as a dark
(but still visible) color #17.

Switch c0/c1 to int16_t so negative values pass through intact,
then map negative c1 -> 0 (black) in screen-palette mode. The
cart's debug text now displays as black, effectively invisible
against the typical dark game background.

Inferred from cart-author behavior (oblivion_eve title cart's
_update60 has print".!5f100..7....6>8.3?.." right after a
pal(1,-15,1)). Not verified against a specific PICO-8 doc.

+++ src/pico8/vm.h
@@ -263,7 +263,7 @@ private:
     void api_mset(int16_t x, int16_t y, uint8_t n);
     void api_oval(int16_t x0, int16_t y0, int16_t x1, int16_t y1, opt<fix32> c);
     void api_ovalfill(int16_t x0, int16_t y0, int16_t x1, int16_t y1, opt<fix32> c);
-    opt<uint8_t> api_private_pal(opt<uint8_t> c0, opt<uint8_t> c1, uint8_t p);
+    opt<uint8_t> api_private_pal(opt<int16_t> c0, opt<int16_t> c1, uint8_t p);
     var<int16_t, bool> api_palt(opt<int16_t> c, opt<bool> t);
     fix32 api_pget(int16_t x, int16_t y);
     void api_pset(int16_t x, int16_t y, opt<fix32> c);
